+++ internal/argocd/client.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/ntmggr/srv-status/internal/kube"
+	"github.com/ntmggr/k8s-status/internal/kube"
 )
 
 const (